Use `SELECTED_ADDRESS_CHANGED` state instead of cached address
The `unconnected-account` reducer no longer caches the selected address
when the alert was opened. Instead a new event was added for when the
selected address changes (`SELECTED_ADDRESS_CHANGED`). Now that we know
exactly when the selected address changes, we don't need to cache the
original account for comparison anymore.